+++ src/modules/users/routes/roles.routes.ts
@@ -13,6 +13,7 @@ const rolesRoute = Router();
 rolesRoute.use((req, res, next) => usersMiddleware.ensureTokenExists(req, res, next));
 rolesRoute.post(
     "", 
+    sharedMiddlewares.ensurePermission("CAN_CREATE_ROLE"),
     sharedMiddlewares.validateSchema(schemas.roles.request),
     (req, res) => rolesController.create(req, res)
 );
@@ -27,12 +28,14 @@ rolesRoute.get(
 );
 rolesRoute.patch(
     "/:id",
+    sharedMiddlewares.ensurePermission("CAN_UPDATE_ROLE"),
     sharedMiddlewares.validateSchema(schemas.roles.update),
     (req, res, next) => rolesMiddleware.ensureRolesIdExists(req, res, next),
     (req, res) => rolesController.updateById(req, res)
 );
 rolesRoute.delete(
     "/:id", 
+    sharedMiddlewares.ensurePermission("CAN_DELETE_ROLE"),
     (req, res, next) => rolesMiddleware.ensureRolesIdExists(req, res, next),
     (req, res) => rolesController.deleteById(req, res)
 );
